Improved indent level caching.
Added support for formatted display of huge queries.

``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22631.3296)
Intel Core i7-10700 CPU 2.90GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=8.0.202
  [Host]     : .NET 6.0.28 (6.0.2824.12007), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.28 (6.0.2824.12007), X64 RyuJIT AVX2


```

### before
|                Method |        Mean |     Error |    StdDev |
|---------------------- |------------:|----------:|----------:|
|          SqModelParse | 1,313.70 μs | 12.967 μs | 12.129 μs |
|    SqModelCommandText |    35.07 μs |  0.254 μs |  0.237 μs |
|      CarbunqlDeepCopy |   110.15 μs |  0.647 μs |  0.573 μs |
|         CarbunqlParse | 1,338.75 μs |  5.366 μs |  5.019 μs |
| CarbunqlToOneLineText |   118.19 μs |  0.398 μs |  0.373 μs |
|        CarbunqlToText |   519.86 μs |  1.930 μs |  1.711 μs |

### after
|                Method |        Mean |     Error |    StdDev |
|---------------------- |------------:|----------:|----------:|
|          SqModelParse | 1,307.94 μs |  3.650 μs |  3.414 μs |
|    SqModelCommandText |    35.88 μs |  0.359 μs |  0.300 μs |
|      CarbunqlDeepCopy |   109.56 μs |  0.624 μs |  0.583 μs |
|         CarbunqlParse | 1,372.50 μs | 12.326 μs | 11.530 μs |
| CarbunqlToOneLineText |   118.16 μs |  0.722 μs |  0.676 μs |
|        CarbunqlToText |   477.57 μs |  2.537 μs |  2.373 μs |